[17] (chore on ratp-wrapper) Moved services into their own folder + better `DEV.md` file.
