+++ DEV.md
@@ -6,7 +6,8 @@ This project uses `sveltekit-i18n` for handling translations and internationaliz
 
 ### Setup
 
-The internationalization system is configured in `src/lib/translations/index.ts`. The configuration includes:
+The internationalization system is configured in [`src/lib/translations/index.ts`](src/lib/translations/index.ts). The
+configuration includes:
 - Available languages (en, fr)
 - Translation loaders for each language
 - Language names in their native form
@@ -39,7 +40,7 @@ To use translations in your Svelte components:
 
 #### 2. Language Selection
 
-The language selector is implemented in the layout component (`src/routes/+layout.svelte`):
+The language selector is implemented in the layout component ([`src/routes/+layout.svelte`](src/routes/+layout.svelte)):
 
 ```svelte
 <script lang="ts">
@@ -56,26 +57,26 @@ The language selector is implemented in the layout component (`src/routes/+layou
 #### 3. Adding New Translations
 
 1. Add new translation keys to the language-specific content files:
-   - `src/lib/translations/en/content.json` for English
-   - `src/lib/translations/fr/content.json` for French
+   - [`src/lib/translations/en/content.json`](src/lib/translations/en/content.json) for English
+   - [`src/lib/translations/fr/content.json`](src/lib/translations/fr/content.json) for French
 
-2. Use nested objects for organizing translations:
-```json
-{
-  "section": {
-    "key": "Translation"
-  }
-}
-```
+2. Use nested objects for organizing translations
+   ```json
+   {
+     "section": {
+       "key": "Translation"
+     }
+   }
+   ```
 
 3. Access nested translations using dot notation:
-```svelte
-{$t('section.key')}
-```
+   ```svelte
+   {$t('section.key')}
+   ```
 
 #### 4. Default Language
 
-The default language is set to French in `src/routes/+layout.ts`:
+The default language is set to French in [`src/routes/+layout.ts`](src/routes/+layout.ts):
 ```typescript
 const initLocale = 'fr';
 ```
@@ -91,10 +92,10 @@ const initLocale = 'fr';
 
 To add a new language:
 
-1. Create a new language directory in `src/lib/translations/` (e.g., `es/` for Spanish)
+1. Create a new language directory in [`src/lib/translations/`](src/lib/translations) (e.g., `es/` for Spanish)
 2. Add the language content file (e.g., `content.json`)
 3. Add the language name to `lang.json`
-4. Update the configuration in `index.ts` to include the new language
+4. Update the configuration in [`index.ts`](src/lib/translations/index.ts) to include the new language
 
 Example for adding Spanish:
 ```typescript
@@ -114,4 +115,67 @@ export const config: Config = {
     },
   ],
 };
+```
+
+## RATP Services
+
+The project includes a service layer for interacting with the RATP API. The implementation is located in
+[`src/lib/utils/ratpServices.ts`](src/lib/services/ratpServices.ts).
+
+### Features
+
+- Singleton pattern for API client management
+- Environment-based API key configuration
+- Type-safe API responses using Zod schemas
+- Support for both v1 and v2 API endpoints
+
+### Usage Example
+
+A small proof of concept (POC) is located in [`scripts/ratp_scripts/index.ts`](scripts/ratp_scripts/index.ts) to
+demonstrate the usage of RATP services.
+
+## Google Services
+
+The project includes a service layer for interacting with the Google Maps API. The implementation is located in
+[`src/lib/utils/googleServices.ts`](src/lib/services/googleServices.ts).
+
+### Features
+
+- Singleton pattern for API client management
+- Environment-based API key configuration
+- Type-safe API responses using Google Maps Services JS types
+- Optimized place details requests
+
+### Available Methods
+
+- `placeAutocomplete`: Get place suggestions based on user input
+- `placeDetails`: Get detailed information about a specific place (optimized for geometry data)
+
+### Usage Example
+
+A small POC is located in [`scripts/google_scripts/index.ts`](scripts/google_scripts/index.ts) to demonstrate the usage
+of Google services.
+
+## Input Debouncing
+
+A custom Svelte action for debouncing input events is available in `src/lib/actions/debounce.ts`.
+
+### Usage Example
+
+```svelte
+<script lang="ts">
+  import { debounce } from '$lib/actions/debounce';
+  
+  let inputValue = '';
+  
+  function handleInput(value: string) {
+    console.log('Input value:', value);
+  }
+</script>
+
+<input 
+  type="text" 
+  bind:value={inputValue} 
+  use:debounce={{ delay: 500, callback: handleInput }} 
+/>
 ```
